+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - **[NEW]** Introduce `LayoutNormalizer` with legacy global layout overrides. Not activated by default.
+- **[NEW]** Add Bus Ride Details story page
 - [...]
 
 # v21.1.0 (19/02/2020)